apply review suggestions
