--- zrml/hybrid-router/src/tests/trade.rs
@@ -1,159 +1,0 @@
-// Copyright 2024 Forecasting Technologies LTD.
-//
-// This file is part of Zeitgeist.
-//
-// Zeitgeist is free software: you can redistribute it and/or modify it
-// under the terms of the GNU General Public License as published by the
-// Free Software Foundation, either version 3 of the License, or (at
-// your option) any later version.
-//
-// Zeitgeist is distributed in the hope that it will be useful, but
-// WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
-// General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with Zeitgeist. If not, see <https://www.gnu.org/licenses/>.
-
-#![cfg(test)]
-
-use super::*;
-
-/// These tests are for shared logic of the buy and sell functions.
-
-#[test]
-fn trade_emits_event() {
-    ExtBuilder::default().build().execute_with(|| {
-        let market_id = 0;
-        let mut market = market_mock::<Runtime>(MARKET_CREATOR);
-        let required_asset_count = match &market.market_type {
-            MarketType::Scalar(_) => panic!("Categorical market type is expected!"),
-            MarketType::Categorical(categories) => *categories,
-        };
-        market.status = MarketStatus::Active;
-        Markets::<Runtime>::insert(market_id, market);
-
-        let asset_count = required_asset_count;
-        let asset = Asset::CategoricalOutcome(market_id, 0);
-        let amount_in = 10 * BASE;
-        let max_price = (BASE / 2).saturated_into::<BalanceOf<Runtime>>();
-        let orders = vec![];
-        let strategy = Strategy::LimitOrder;
-        assert_ok!(HybridRouter::buy(
-            RuntimeOrigin::signed(ALICE),
-            market_id,
-            asset_count,
-            asset,
-            amount_in,
-            max_price,
-            orders,
-            strategy,
-        ));
-
-        System::assert_last_event(
-            Event::<Runtime>::HybridRouterBuyExecuted {
-                who: ALICE,
-                market_id,
-                asset,
-                amount_in,
-                max_price,
-            }
-            .into(),
-        );
-    });
-}
-
-#[test]
-fn trade_fails_if_asset_count_mismatch() {
-    ExtBuilder::default().build().execute_with(|| {
-        let market_id = 0;
-        let mut market = market_mock::<Runtime>(MARKET_CREATOR);
-        let required_asset_count = match &market.market_type {
-            MarketType::Scalar(_) => panic!("Categorical market type is expected!"),
-            MarketType::Categorical(categories) => *categories,
-        };
-        market.status = MarketStatus::Active;
-        Markets::<Runtime>::insert(market_id, market);
-
-        let asset_count = 2;
-        assert_ne!(required_asset_count, asset_count);
-        let asset = Asset::CategoricalOutcome(market_id, 0);
-        let amount = 2 * BASE;
-        let max_price = (BASE / 2).saturated_into::<BalanceOf<Runtime>>();
-        let orders = vec![];
-        let strategy = Strategy::ImmediateOrCancel;
-        assert_noop!(
-            HybridRouter::buy(
-                RuntimeOrigin::signed(ALICE),
-                market_id,
-                asset_count,
-                asset,
-                amount,
-                max_price,
-                orders,
-                strategy,
-            ),
-            Error::<Runtime>::AssetCountMismatch
-        );
-    });
-}
-
-#[test]
-fn trade_fails_if_cancel_strategy_applied() {
-    ExtBuilder::default().build().execute_with(|| {
-        let market_id = 0;
-        let mut market = market_mock::<Runtime>(MARKET_CREATOR);
-        let required_asset_count = match &market.market_type {
-            MarketType::Scalar(_) => panic!("Categorical market type is expected!"),
-            MarketType::Categorical(categories) => *categories,
-        };
-        market.status = MarketStatus::Active;
-        Markets::<Runtime>::insert(market_id, market);
-
-        let asset_count = required_asset_count;
-        let asset = Asset::CategoricalOutcome(market_id, 0);
-        let amount = 10 * BASE;
-        let max_price = (BASE / 2).saturated_into::<BalanceOf<Runtime>>();
-        let orders = vec![];
-        let strategy = Strategy::ImmediateOrCancel;
-        assert_noop!(
-            HybridRouter::buy(
-                RuntimeOrigin::signed(ALICE),
-                market_id,
-                asset_count,
-                asset,
-                amount,
-                max_price,
-                orders,
-                strategy,
-            ),
-            Error::<Runtime>::CancelStrategyApplied
-        );
-    });
-}
-
-#[test]
-fn trade_fails_if_market_does_not_exist() {
-    ExtBuilder::default().build().execute_with(|| {
-        let market_id = 0;
-        let asset_count = 2;
-        let asset = Asset::CategoricalOutcome(market_id, 0);
-        let amount = 10 * BASE;
-        let max_price = (BASE / 2).saturated_into::<BalanceOf<Runtime>>();
-        let orders = vec![];
-        let strategy = Strategy::ImmediateOrCancel;
-        assert_noop!(
-            HybridRouter::buy(
-                RuntimeOrigin::signed(ALICE),
-                market_id,
-                asset_count,
-                asset,
-                amount,
-                max_price,
-                orders,
-                strategy,
-            ),
-            MError::<Runtime>::MarketDoesNotExist
-        );
-    });
-}
